refactor(Products): change type of id to long

+++ src/main/java/gift/model/Products.java
@@ -10,7 +10,7 @@ import java.util.NoSuchElementException;
 public class Products {
 
     // id를 통한 빠른 탐색을 위해 HashMap을 사용.
-    private Map<Integer, Product> products;
+    private Map<Long, Product> products;
 
     // 생성자
     public Products() {
@@ -19,7 +19,7 @@ public class Products {
 
     // map에 제품을 넣는 메서드
     public Product put(Product product) {
-        int id = product.getId();
+        long id = product.getId();
         verifyProductAlreadyExist(id);
         products.put(id, product);
 
@@ -28,7 +28,7 @@ public class Products {
     }
 
     // id에 맞는 상품 조회
-    public Product getProduct(int id) {
+    public Product getProduct(long id) {
         // 도메인 규칙 검사
         verifyProductExist(id);
         // 멀쩡하면 불러오기
@@ -46,7 +46,7 @@ public class Products {
     }
 
     // id에 맞는 데이터를 삭제
-    public void deleteProduct(int id) {
+    public void deleteProduct(long id) {
         // 우선 해당하는 id가 있는지부터 검사
         verifyProductExist(id);
 
@@ -62,7 +62,7 @@ public class Products {
     // 상품 수정하기 (patch)
     public Product updateProduct(int targetId, Product product) {
         // 이미 존재하는 id로 수정하면 안됨(자신 포함)
-        int id = product.getId();
+        long id = product.getId();
         verifyProductAlreadyExist(id);
 
         // products에서의 검증이 끝났으면 product에서 검증 후 변경
@@ -79,14 +79,14 @@ public class Products {
     }
 
     // 도메인 규칙: 삽입 시에는 id가 중복되지 않아야 한다.
-    private void verifyProductAlreadyExist(int id) {
+    private void verifyProductAlreadyExist(long id) {
         if (products.containsKey(id)) {
             throw new ValidationException("id가 중복됩니다.");
         }
     }
 
     // 도메인 규칙: 조회 시에는 id가 존재해야 한다.
-    private void verifyProductExist(int id) {
+    private void verifyProductExist(long id) {
         if (!products.containsKey(id)) {
             throw new NoSuchElementException("해당 id를 가진 제품이 존재하지 않습니다.");
         }
